refactor: embeding
