+++ apps/commons/mixins.py
@@ -408,3 +408,24 @@ class HasMultipleIDs:
             if self.get_id_field_name(slug) != "slug":
                 slug = f"{self.slug_prefix}-{slug}"
         return slug
+
+
+
+class HasEmbending:
+    def vectorize(self):
+        if not getattr(self, "embedding", None):
+            model_embedding = type(self.embedding)
+            self.embedding = model_embedding(item=self)
+            self.embedding.save()
+        self.embedding.vectorize()
+
+    def similars(self, threshold: float = 0.15) -> QuerySet[Self]:
+        """return similars documents"""
+        if getattr(self, "embedding", None):
+            vector = self.embedding.embedding
+            model_embedding = type(self.embedding)
+            queryset = type(self).objects.all()
+            return model_embedding.vector_search(vector, queryset, threshold).exclude(
+                pk=self.pk
+            )
+        return type(self).objects.all()
